+++ application/controllers/Booking_confirmation.php
@@ -63,6 +63,7 @@ class Booking_confirmation extends EA_Controller
             'matomo_analytics_url' => setting('matomo_analytics_url'),
             'matomo_analytics_site_id' => setting('matomo_analytics_site_id'),
             'add_to_google_url' => $add_to_google_url,
+            'display_add_to_google_calendar' => setting('display_add_to_google_calendar', '1'),
         ]);
 
         $this->load->view('pages/booking_confirmation');