+++ application/views/pages/booking_confirmation.php
@@ -24,12 +24,14 @@
                 <?= lang('go_to_booking_page') ?>
             </a>
 
+            <?php if (vars('display_add_to_google_calendar') === '1'): ?>
             <a href="<?= vars(
                 'add_to_google_url',
             ) ?>" id="add-to-google-calendar" class="btn btn-outline-primary px-4 py-2" target="_blank">
                 <i class="fab fa-google me-2"></i>
                 <?= lang('add_to_google_calendar') ?>
             </a>
+            <?php endif; ?>
         </div>
     </div>
 </div>